Source files are now organized
