+++ src/load_map.c
