+++ src/movement_utils.c
